+++ src/main/java/by/clevertec/service/PersonWithRangeEvacuation.java
@@ -1,7 +1,8 @@
-package by.clevertec.model;
+package by.clevertec.service;
 
 
-import by.clevertec.util.EvacuationRank;
+import by.clevertec.model.Person;
+import by.clevertec.service.EvacuationRank;
 import by.clevertec.util.TaskUtil;
 import lombok.AllArgsConstructor;
 import lombok.Data;
